Add worm support
We want to add the first version of worm support. The implementation
would prevent users from modifying, touching, or removing of files
with the worm attribute set.

Note in this version; we also prevent the root user from modifying
the expiration time once one is set. The user would need to wait
until expiration prior to being able to modify the attribute or file.
This was done by design/request.

Signed-off-by: Bryant G. Duffy-Ly <bduffyly@versity.com>